implement setStone, removeStone
Realized that the simple formula for updating a group's liberties (new liberties = old liberties - 1 + liberties of new stone) doesn't work around corners. I am considering using a Set-based implementation, the alternative would seem to be to check for existing groups in a 3x3 grid with the new stone at its center, though that would still not account for the merging of formerly separate groups that share some liberties, which could be even more complicated.