Add move number capability
Completed transition to using StackPane inside the GridPane Cells. Contemplating extending StackPane to be able to address all components of a board cell in a readable way (currently requires knowing the index).

BoardPane can now show move numbers, which it currently receives from the Board in the StoneSetEvent (this responsibility might have to be handed over to the Game and its GameEvents).