have StoneGroups keep track of their pointers
Sadly, this is no longer O(1) complexity, but it is still at worst O(n) and in most real-life use-cases would have less than linear complexity, as the number of pointers used by a group will usually be lower than the number of positions it occupies.

What's left to do is to untangle some of the now unnecessary complexity in the source code.