use AnchorPane instead of BoardPane
BoardPane didn't seem to let me force the coordinates to stick directly to the board, so I used AnchorPane instead, which, after hours of tinkering, I could finally get to align the coordinates (only the vertical ones for now) properly. Sometimes the coordinates are a single pixel too close or too far away, but that's hardly noticeable during actual usage and, in my opinion, worthwile for actually having a neatly packaged solution that has coordinates which adhere.

I will have to clean this up, of course.