Make Ruleset compile-time based
As a standard Ruleset won't change during gameplay, it seems better to hard-code its properties in its own class, rather than set it up at runtime. The exception to this would be user-defined rulesets, which we didn't promise, but if we wanted to implement them, the UserDefinableRuleset class shows how.