simplify capture check
The simplified condition is only ever reached if suicide was not detected or permitted. Thus, FINAL_KILL_ANOTHER would never be false if we even got here with the JapaneseRuleset, and if we used the NewZealandRuleset and committed suicide, any neighbouring groups of the same color should be eliminated in this collective suicide.