add debug in BoardPane
I'm looking for an issue where if you undo the first move, all handicap stones disappear. I think I've found a clue: For some reason, undoing the first move sends a NEW_GAME Gamecommand to the boardpane. This will be because the old GameCOmmand of Game is restored when undoing.