fix bug
It turns out that, while the collections returned by Collections.unmodifiableX cannot be changed via the variable itself, they remain a mirror image of the original. Thus, if the original changes, so does the unmodifiable version. The X.copyOf methods should therefore be used instead, as these provide unmodifiable copies.