make History more uniform
History now has an AbstractSaveToken for its beginning, which is useful for demonstrating intent (i.e., one can now be sure that a HistoryNode is the beginning of history and not just an invalid HistoryNode with a save token of null).

Additionally, I realised that having TERMINATOR be static would be problematic in the event of having several instances of History, for obvious reasons.