remove game.SaveGame() and game.loadGame()
Seeing as it's impossible to do anything undesirable by getting access to the Game's FileHandler() and as Game's interface is already quite large, these methods, which simply pass a return value, seem superfluous. This is especially true as the FileHandler has to be accessible to begin with to obtain the current file name, which definitely does not seem reasonable for Game's interface to include.