add factory method to UndoableCommand
If all an UndoableCommand does is update a value (which is increasingly likely due to the increasing modularisation of the UndoableCommands), then an anonymous UndoableCommand just for this is ridiculously huge - 11 lines for what would take a single line without the UndoableCommand. To mitigate this, we now have to factory method UndoableCommand.updateValue() which creates such a simple Undoablecommand.